requirements.txt
langchain>=0.1.0
langchain-core>=0.1.0
langchain-community>=0.0.10
langchain-experimental>=0.0.40
langchain-openai>=0.0.5
langchain-text-splitters>=0.0.1
langchain-chroma>=0.0.1
langchain-anthropic>=0.0.1
langchain-huggingface>=0.0.1
langgraph>=0.0.20
chromadb>=0.4.18
openai>=1.3.0
anthropic>=0.5.0
transformers>=4.35.0
sentence-transformers>=2.2.2
beautifulsoup4>=4.12.2
selenium>=4.15.0
requests>=2.31.0
arxiv>=1.4.8
scholarly>=1.7.11
flask>=2.3.3
fastapi>=0.104.1
uvicorn>=0.24.0
pydantic>=2.4.2
python-dotenv>=1.0.0
numpy>=1.26.0
pandas>=2.1.1
matplotlib>=3.8.0
seaborn>=0.13.0
plotly>=5.18.0
nbformat>=5.9.2
nbconvert>=7.9.2
jupyter>=1.0.0
ipykernel>=6.25.2
redis>=5.0.1
aiohttp>=3.8.6
pytest>=7.4.3
